🚀 Refactor login controller
